dnsResolver/requirements.txt
cffi==1.14.6
cryptography==3.4.8
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
fonttools==4.37.2
kiwisolver==1.4.4
matplotlib==3.5.0
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
